refactor: extract Calendar component from TodayPage
